0.1.1/requirements.txt
# Requirements for Autoimmune Treatment Recommendation Application
# Core dependencies
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.19
pydantic==2.0.3
pytest==7.4.0

# Data processing
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0

# API consumption for clinical trials
requests==2.31.0
beautifulsoup4==4.12.2

# Integration with LLM services
openai==0.27.8

# Web interface
streamlit==1.24.1
plotly==5.15.0

# Machine Learning
xgboost==2.0.0
optuna==3.2.0
kaggle==1.5.16
matplotlib==3.7.2
seaborn==0.12.2
shap==0.42.1 